requirements.txt
asgiref==3.8.1
Babel==2.15.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
fonttools==4.53.1
gunicorn==22.0.0
html5lib==1.1
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tinycss2==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3
